+++ src/components/music/form/WaveEditorForm.tsx
@@ -76,7 +76,9 @@ export const WaveEditorForm = ({ waveId, onChange }: WaveEditorFormProps) => {
       return;
     }
 
-    const defaultColor = themeContext?.colors.highlight ?? "black";
+    const defaultColor = themeContext?.colors.tracker.wave ?? "#fff";
+    const backgroundColor =
+      themeContext?.colors.tracker.waveBackground ?? "#000";
 
     const getLayout = () => {
       const rect = canvas.getBoundingClientRect();
@@ -115,7 +117,7 @@ export const WaveEditorForm = ({ waveId, onChange }: WaveEditorFormProps) => {
     const clear = () => {
       const { width, height } = getLayout();
       ctx.clearRect(0, 0, width, height);
-      ctx.fillStyle = "#000";
+      ctx.fillStyle = backgroundColor;
       ctx.fillRect(0, 0, width, height);
     };
 
@@ -280,7 +282,13 @@ export const WaveEditorForm = ({ waveId, onChange }: WaveEditorFormProps) => {
       window.removeEventListener("touchcancel", handleTouchEnd);
       window.removeEventListener("resize", handleResize);
     };
-  }, [onEditWave, songWave, themeContext?.colors.highlight, wavesLength]);
+  }, [
+    onEditWave,
+    songWave,
+    themeContext?.colors.tracker.wave,
+    themeContext?.colors.tracker.waveBackground,
+    wavesLength,
+  ]);
 
   return (
     <>