+++ src/components/ui/theme/lightTheme.ts
@@ -158,6 +158,9 @@ const lightTheme: ThemeInterface = {
         border: "#3b3a3b66",
       },
       sharpBackground: "#3aa1d511",
+      wave: "#c92c61",
+      waveGrid: "#333333",
+      waveBackground: "#000000",
     },
     prefab: {
       background: "#03a9f4",